refactor: split html diff report internals
